+++ src/components/CalloutCard/CalloutCard.docs.mdx
@@ -1,14 +1,14 @@
 import { Story, Preview } from '@storybook/addon-docs/blocks';
-import InlineWrapper from './InlineWrapper';
+import CalloutCard from './CalloutCard';
 
-# InlineWrapper
+# Callout Card
 
 A stylized box that brings attention to a piece of text.
 
 ## Usage
 
 ```js
-import { InlineWrapper } from '@quartz/interface/src/components';
+import { CalloutCard } from '@quartz/interface/src/components';
 ```
 
 ## Props
@@ -19,5 +19,5 @@ This component accepts children as props and wraps them with a stylized containe
 
 ### Default
 <Preview>
-	<Story id="inlinewrapper--default" />
+	<Story id="calloutcard--default" />
 </Preview>
