+++ src/components/index.js
@@ -1,5 +1,5 @@
 export { default as Button, ButtonLink } from './Button/Button';
-export { default as InlineWrapper } from './InlineWrapper/InlineWrapper';
+export { default as CalloutCard } from './CalloutCard/CalloutCard';
 export { default as Checkbox } from './Checkbox/Checkbox';
 export { default as Link } from './Link/Link';
 export { default as RadioButton } from './RadioButton/RadioButton';
